:bricks: Building Narrative Changes' Scenes
- Created Narrative Changes' Scenes
- Polished Butterflies' and Dragonflies' Animations
- Changed Waterfall Material
- Polished Animal NavMesh Floor Spacing
- Painted Terrain Textures, Grass and Leaf Litter
- Added Notebook Related Materials and Post-it Icons
- Changed Animal Names
- Added Game Logo
- Imported New Cinematic and Credits Videos
- Adjusted Notification's and Arpeggios's Volumes

+++ ProyectoAbueloUnity/Assets/Core/Scripts/UI/NotebookCursorTarget/SoundSliderCT.cs
@@ -8,6 +8,6 @@ public class SoundSliderCT : SliderCT
     {
         base.SetLevel(level);
 
-        SettingsManager.Instance.SetSFXVolume(level / _levels);
+        SettingsManager.Instance.SetSFXVolume((float)level / (float)_levels);
     }
 }
